2023/04/10
1. 어제는 약속이 있어 작업을 하지 못했다.

2. 기존에는 서블릿response에서 헤더를 가져와 토큰을 얻는 것으로 알고 있었는데 이상하게 테스트 중 에러가 발생하여 로그를 찍어보니 헤더가 null로 나왔고 따라서 response에서 헤더를 가져오는 것이 아니라는 것을 을 알게되었으며.

3. requestHeader 어노테이션을 사용하여 헤더를 가져오는 것을 알게되었다.

4. 2,3 때문에 시간을 많이 소비하여 작업을 많이 진행하지 못했다.

5. 내일은 남은 작업을 완료하고 클라이언트단을 작업할 예정이다.

+++ src/main/java/Baetube_backEnd/controller/RestPlaylistController.java
@@ -15,15 +15,18 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 
 import Baetube_backEnd.ErrorResponse;
+import Baetube_backEnd.dto.Channel;
 import Baetube_backEnd.dto.Playlist;
 import Baetube_backEnd.dto.PlaylistItem;
 import Baetube_backEnd.dto.Video;
 import Baetube_backEnd.exception.DuplicatePlaylistItemException;
 import Baetube_backEnd.exception.NullPlaylistException;
 import Baetube_backEnd.exception.NullPlaylistItemException;
+import Baetube_backEnd.service.jwt.JwtTokenDataExtractService;
 import Baetube_backEnd.service.playlist.PlaylistChannelService;
 import Baetube_backEnd.service.playlist.PlaylistDeleteItemService;
 import Baetube_backEnd.service.playlist.PlaylistDeleteLikeVideoService;
@@ -55,6 +58,8 @@ public class RestPlaylistController
 	private PlaylistInsertLikeVideoService playlistInsertLikeVideoService;
 	@Autowired
 	private PlaylistDeleteLikeVideoService playlistDeleteLikeVideoService;
+	@Autowired
+	private JwtTokenDataExtractService jwtTokenDataExtractService;
 	
 	@GetMapping("/api/playlist/channel/{channelId}")
 	//@ExceptionHandler(MethodArgumentNotValidException.class)
@@ -62,7 +67,6 @@ public class RestPlaylistController
 	{                                             
 		try
 		{
-			System.out.println("요청이 도착했습니다.");
 			List<Playlist> playlistList = playlistChannelService.select(channelId);
 			return ResponseEntity.status(HttpStatus.OK).body(playlistList);
 		} 
@@ -73,21 +77,15 @@ public class RestPlaylistController
 		}
 	}
 	
-	@PostMapping("/api/playlist/insert")
+	@GetMapping("/api/storage/channel/{channelSequence}")
 	//@ExceptionHandler(MethodArgumentNotValidException.class)
-	public ResponseEntity<Object> insertPlaylist(@RequestBody @Valid Playlist request, Errors errors, HttpServletResponse response) throws IOException
-	{
-		if(errors.hasErrors())
-		{
-			String errorCodes = errors.getAllErrors().stream().map(error -> error.getCodes()[0]).collect(Collectors.joining(","));
-			
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ErrorResponse("errorCodes = " + errorCodes));
-		}
-		                                                 
+	public ResponseEntity<Object> selectChannelStorage(@RequestHeader("Authorization") String bearerToken, @PathVariable Integer channelSequence, HttpServletResponse response) throws IOException
+	{                                             
 		try
 		{
-			Integer playlistId = playlistInsertService.insert(request);
-			return ResponseEntity.status(HttpStatus.OK).body(playlistId);
+			Channel channel = jwtTokenDataExtractService.getChannelData(bearerToken, channelSequence);
+			List<Playlist> playlistList = playlistChannelService.select(channel.getChannelId());
+			return ResponseEntity.status(HttpStatus.OK).body(playlistList);
 		} 
 		catch (NullPlaylistException e)
 		{
@@ -96,9 +94,9 @@ public class RestPlaylistController
 		}
 	}
 	
-	@PostMapping("/api/playlist/delete")
+	@PostMapping("/api/playlist/insert/{channelSequence}")
 	//@ExceptionHandler(MethodArgumentNotValidException.class)
-	public ResponseEntity<Object> deletePlaylist(@RequestBody @Valid Playlist request, Errors errors, HttpServletResponse response) throws IOException
+	public ResponseEntity<Object> insertPlaylist(@RequestHeader("Authorization") String bearerToken, @PathVariable Integer channelSequence, @RequestBody @Valid Playlist request, Errors errors, HttpServletResponse response) throws IOException
 	{
 		if(errors.hasErrors())
 		{
@@ -109,9 +107,32 @@ public class RestPlaylistController
 		                                                 
 		try
 		{
-			System.out.println("요청이 도착했습니다.");
-			System.out.println("playlistId : " + request.getPlaylistId());
-			System.out.println("channelId : " + request.getChannelId());
+			Channel channel = jwtTokenDataExtractService.getChannelData(bearerToken, channelSequence);			
+			request.setChannelId(channel.getChannelId());
+			Integer playlistId = playlistInsertService.insert(request);
+			return ResponseEntity.status(HttpStatus.OK).body(playlistId);
+		} 
+		catch (NullPlaylistException e)
+		{
+			return ResponseEntity.status(HttpStatus.CONFLICT).build();
+		}
+	}
+	
+	@PostMapping("/api/playlist/delete/{channelSequence}")
+	//@ExceptionHandler(MethodArgumentNotValidException.class)
+	public ResponseEntity<Object> deletePlaylist(@RequestHeader("Authorization") String bearerToken, @PathVariable Integer channelSequence, @RequestBody @Valid Playlist request, Errors errors, HttpServletResponse response) throws IOException
+	{
+		if(errors.hasErrors())
+		{
+			String errorCodes = errors.getAllErrors().stream().map(error -> error.getCodes()[0]).collect(Collectors.joining(","));
+			
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ErrorResponse("errorCodes = " + errorCodes));
+		}
+		                                                 
+		try
+		{
+			Channel channel = jwtTokenDataExtractService.getChannelData(bearerToken, channelSequence);			
+			request.setChannelId(channel.getChannelId());
 			playlistDeleteService.delete(request);
 			return ResponseEntity.status(HttpStatus.OK).build();
 		} 
@@ -135,10 +156,6 @@ public class RestPlaylistController
 		                                                 
 		try
 		{
-			System.out.println("요청이 도착했습니다.");
-			System.out.println("playlistId : " + request.getPlaylistId());
-			System.out.println("channelId : " + request.getChannelId());
-			System.out.println("thumbnail : " + request.getThumbnail());
 			playlistUpdateService.update(request);
 			return ResponseEntity.status(HttpStatus.OK).build();
 		} 
@@ -196,12 +213,12 @@ public class RestPlaylistController
 	}
 	
 	@GetMapping("/api/playlist/item/{playlistId}")
-	public ResponseEntity<Object> selectPlaylistItem(@PathVariable Integer channelId, HttpServletResponse response) throws IOException
+	public ResponseEntity<Object> selectPlaylistItem(@PathVariable Integer playlistId, HttpServletResponse response) throws IOException
 	{                                             
 		try
 		{
 			System.out.println("요청이 도착했습니다.");
-			List<Playlist> playlistList = playlistChannelService.select(channelId);
+			List<Playlist> playlistList = playlistChannelService.select(playlistId);
 			return ResponseEntity.status(HttpStatus.OK).body(playlistList);
 		} 
 		catch (NullPlaylistException e)
@@ -234,9 +251,9 @@ public class RestPlaylistController
 		}
 	}
 	
-	@PostMapping("/api/playlist/item/insert/like")
+	@PostMapping("/api/playlist/item/insert/like/{channelSequence}")
 	//@ExceptionHandler(MethodArgumentNotValidException.class)
-	public ResponseEntity<Object> insertLikeVideo(@RequestBody @Valid Video request, Errors errors, HttpServletResponse response) throws IOException
+	public ResponseEntity<Object> insertLikeVideo(@RequestHeader("Authorization") String bearerToken, @PathVariable Integer channelSequence, @RequestBody @Valid Video request, Errors errors, HttpServletResponse response) throws IOException
 	{
 		if(errors.hasErrors())
 		{
@@ -247,6 +264,8 @@ public class RestPlaylistController
 		                                                 
 		try
 		{
+			Channel channel = jwtTokenDataExtractService.getChannelData(bearerToken, channelSequence);
+			request.setChannelId(channel.getChannelId());
 			playlistInsertLikeVideoService.insertLikeVideo(request);
 			return ResponseEntity.status(HttpStatus.OK).build();
 		} 
@@ -257,9 +276,9 @@ public class RestPlaylistController
 		}
 	}
 	
-	@PostMapping("/api/playlist/item/delete/like")
+	@PostMapping("/api/playlist/item/delete/like/{channelSequence}")
 	//@ExceptionHandler(MethodArgumentNotValidException.class)
-	public ResponseEntity<Object> deleteLikeVideo(@RequestBody @Valid Video request, Errors errors, HttpServletResponse response) throws IOException
+	public ResponseEntity<Object> deleteLikeVideo(@RequestHeader("Authorization") String bearerToken, @PathVariable Integer channelSequence, @RequestBody @Valid Video request, Errors errors, HttpServletResponse response) throws IOException
 	{
 		if(errors.hasErrors())
 		{
@@ -270,6 +289,8 @@ public class RestPlaylistController
 		                                                 
 		try
 		{
+			Channel channel = jwtTokenDataExtractService.getChannelData(bearerToken, channelSequence);
+			request.setChannelId(channel.getChannelId());
 			playlistDeleteLikeVideoService.deleteLikeVideo(request);
 			return ResponseEntity.status(HttpStatus.OK).build();
 		}
